fix: preserve custom entity glow with glow hiders
Linked issue: #279
